requirements.txt
# Base
pillow==8.3.1
ipdb==0.13.9

# Django
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1

# Django Rest Framework
djangorestframework==3.12.4

# Authentication
PyJWT==2.1.0

# Django Filter
django-filter==2.4.0

# Postgres
psycopg2==2.9.1

# Celery
amqp==5.0.6
celery==5.1.2
django-celery-beat==2.2.1

# Channels
channels==3.0.4
channels-redis==3.3.0